updated it configuration
