+++ processor/src/it/java/com/ihren/processor/config/KafkaTemplateConfig.java
@@ -1,8 +1,6 @@
 package com.ihren.processor.config;
 
 import com.ihren.processor.model.Transaction;
-import com.ihren.processor.serialization.PersonSerializer;
-import com.ihren.processor.model.Person;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,7 +17,7 @@ public class KafkaTemplateConfig {
     @Value("${spring.kafka.bootstrap-servers}")
     private String bootstrapServers;
 
-    @Bean
+    //@Bean
     public ProducerFactory<String, /*TODO: verify*/ Transaction> producerFactory() {
         Map<String, Object> config = Map.of(
                 ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers,
