+++ processor/src/main/java/com/ihren/processor/config/ObjectMapperConfig.java
